requirements_base.txt
otree>=2.6.0b5
pyyaml
pandas
django_pandas
django_pivot
django-webpack-loader==0.7.0
